Reintegrate tags in scene object component
