--- Source/Core/Editor/UI/TagContainerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b211179509634de4e8060f1426e5840f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
