--- Source/Core/Runtime/SceneObjects/ProcessTagContainer.cs
@@ -1,89 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using UnityEngine;
-
-namespace VRBuilder.Core.SceneObjects
-{
-    /// <summary>
-    /// Monobehavior implementation of <see cref="ITagContainer"/>.
-    /// </summary>
-    [ExecuteInEditMode]
-    public class ProcessTagContainer : MonoBehaviour, ITagContainer
-    {
-        [SerializeField]
-        protected List<string> tags = new List<string>();
-
-        /// <inheritdoc />
-        public IEnumerable<Guid> Tags => tags.Select(tag => Guid.Parse(tag));
-
-        /// <inheritdoc />
-        public bool IsLocked { get; private set; }
-
-        /// <inheritdoc />
-        public event EventHandler<TaggableObjectEventArgs> TagAdded;
-
-        /// <inheritdoc />
-        public event EventHandler<TaggableObjectEventArgs> TagRemoved;
-
-        /// <inheritdoc />
-        public event EventHandler<LockStateChangedEventArgs> Locked;
-
-        /// <inheritdoc />
-        public event EventHandler<LockStateChangedEventArgs> Unlocked;
-
-        /// <inheritdoc />
-        public void AddTag(Guid tag)
-        {
-            if (Tags.Contains(tag) == false)
-            {
-                tags.Add(tag.ToString());
-                TagAdded?.Invoke(this, new TaggableObjectEventArgs(tag.ToString()));
-            }
-        }
-
-        /// <inheritdoc />
-        public bool HasTag(Guid tag)
-        {
-            return Tags.Contains(tag);
-        }
-
-        /// <inheritdoc />
-        public bool RemoveTag(Guid tag)
-        {
-            if(tags.Remove(tag.ToString()))
-            {
-                TagRemoved?.Invoke(this, new TaggableObjectEventArgs(tag.ToString()));
-                return true;
-            }
-
-            return false;
-        }
-
-        /// <inheritdoc />
-        public void SetLocked(bool lockState)
-        {
-            if (IsLocked == lockState)
-            {
-                return;
-            }
-
-            IsLocked = lockState;
-
-            if (IsLocked)
-            {
-                if (Locked != null)
-                {
-                    Locked.Invoke(this, new LockStateChangedEventArgs(IsLocked));
-                }
-            }
-            else
-            {
-                if (Unlocked != null)
-                {
-                    Unlocked.Invoke(this, new LockStateChangedEventArgs(IsLocked));
-                }
-            }
-        }
-    }
-}
