--- Source/Core/Runtime/SceneObjects/ProcessTagContainer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db133809e3af7d945a2fcd24be9d3c46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
