+++ Source/Core/Runtime/SceneObjects/SceneObjectRegistry.cs
@@ -129,7 +129,7 @@ namespace VRBuilder.Core.SceneObjects
         /// <inheritdoc />
         public IEnumerable<ISceneObject> GetByTag(Guid tag)
         {
-            return registeredEntities.Values.Where(entity => entity.TagContainer.HasTag(tag));
+            return registeredEntities.Values.Where(entity => entity as ITagContainer != null && ((ITagContainer)entity).HasTag(tag));
         }
     }
 }
